Move pages files to screens
